refactor(desktop): 将飞书通知和企业微信群通知迁移至 IM 机器人设置页
- NotificationSection: 移除飞书通知开关和企业微信群通知卡片，仅保留桌面通知
- FeishuBotNotificationSection: 新增「任务完成通知」开关，与上下线通告并列显示
- WecomBotSection: 底部增加企业微信群通知配置区（webhook/开关/测试/清除）
- ScheduleFormDialog: goConfigWecomGroup 跳转目标更新为 IM 机器人页
- i18n: 四语新增 settings.feishuBot.sessionNotification 键

这两种 IM 通知在逻辑上属于 IM 通道配置，放在 IM 机器人页各自渠道卡片内更符合直觉。
